+++ tests/Test-Elevation.Tests.ps1
@@ -5,17 +5,17 @@ BeforeAll {
     Import-Module "$PSScriptRoot/../easypeasy.psd1" -Force
 }
 
-Describe 'Test-Elevated' {
+Describe 'Test-Elevation' {
 
     It 'exports the command' {
-        Get-Command Test-Elevated -Module easypeasy | Should -Not -BeNullOrEmpty
+        Get-Command Test-Elevation -Module easypeasy | Should -Not -BeNullOrEmpty
     }
 
     It 'returns a boolean' {
-        Test-Elevated | Should -BeOfType [bool]
+        Test-Elevation | Should -BeOfType [bool]
     }
 
     It 'reports the elevation of the current session' {
-        Test-Elevated | Should -Be $isAdmin
+        Test-Elevation | Should -Be $isAdmin
     }
 }